+++ Employeemanagement-main/springapp/src/main/java/com/examly/springapp/repository/JobDescriptionRepo.java
@@ -6,7 +6,7 @@ import com.examly.springapp.model.JobDescription;
 
 
 
-public interface JobDescriptionRepo extends JpaRepository<JobDescription, Integer> {
+public interface JobDescriptionRepo extends JpaRepository<JobDescription, Long> {
 
     
 }
